Minor Improvements

+++ data-view-remote/src/app/remote-entry/entry.component.ts
@@ -32,34 +32,40 @@ export class RemoteEntryComponent {
   public readonly productService = inject(ProductService);
   public readonly messageService = inject(MessageService);
 
-  onRowEditInit(product: IProduct) {
-    // this.clonedProducts[product.id as string] = { ...product };
-  }
-
   onRowEditSave(product: IProduct) {
-    this.productService.editProduct(product);
-    // if (product.price! > 0) {
-    //   delete this.clonedProducts[product.id as string];
-    //   this.messageService.add({
-    //     severity: 'success',
-    //     summary: 'Success',
-    //     detail: 'Product is updated',
-    //   });
-    // } else {
-    //   this.messageService.add({
-    //     severity: 'error',
-    //     summary: 'Error',
-    //     detail: 'Invalid Price',
-    //   });
-    // }
-  }
-
-  onRowEditCancel(product: IProduct, index: number) {
-    // this.products[index] = this.clonedProducts[product.id as string];
-    // delete this.clonedProducts[product.id as string];
+    if (product) {
+      this.productService.editProduct(product);
+      this.messageService.add({
+        severity: 'success',
+        summary: 'Success',
+        detail: 'Product Has Been Updated',
+      });
+    } else {
+      this.messageService.add({
+        severity: 'error',
+        summary: 'Error',
+        detail: 'Something Went Wrong',
+      });
+    }
   }
 
   onRowEditDelete(product: IProduct, index: number) {
-    this.productService.deleteProduct(product);
+    if (product) {
+      this.productService.deleteProduct(product);
+      this.messageService.add({
+        severity: 'success',
+        summary: 'Success',
+        detail: 'Product Has Been Deleted',
+      });
+    } else {
+      this.messageService.add({
+        severity: 'error',
+        summary: 'Error',
+        detail: 'Something Went Wrong',
+      });
+    }
   }
+
+  onRowEditCancel(product: IProduct, index: number) {}
+  onRowEditInit(product: IProduct) {}
 }
